requirements.txt
langchain>=0.3
langchain-anthropic>=0.3
python-dotenv>=1.0
flask>=3.0
gunicorn>=22.0
